+++ examples/slice-pizza/datadog-app/src/config.js
@@ -1,0 +1,2 @@
+export const PROXY_URL = process.env.REACT_APP_PROXY_URL
+
