+++ examples/slice-pizza/datadog-app/src/types/Order.ts
@@ -1,0 +1,7 @@
+import Pizza from './Pizza'
+
+export default interface Order {
+    items: Pizza[];
+    total: number;
+}
+
