+++ examples/slice-pizza/datadog-app/src/types/Pizza.ts
@@ -1,0 +1,8 @@
+export default interface Pizza {
+    description: string;
+    id: string;
+    name: string;
+    price: number;
+    quantity: number;
+}
+
