+++ examples/slice-pizza/datadog-app/src/types/Token.ts
@@ -1,0 +1,6 @@
+export default interface Token {
+    email: string;
+    id: string;
+    expires: number;
+}
+
