+++ examples/slice-pizza/datadog-app/src/types/User.ts
@@ -1,0 +1,5 @@
+export default interface User {
+    email: string;
+    password: string;
+}
+
